refactor: fix vercel deploy
